third_party/WebKit/LayoutTests/fast/js/Promise-simple-resolve.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

var future = new Promise(function(resolve) {
  setTimeout(function() {
    resolve('hello');
  }, 0);
}).then(function(result) {
  window.result = result;
  shouldBeEqualToString('result', 'hello');
  finishJSTest();
});

</script>
</body>
</html>
